refactor(types): Put cert ref into submodule
