+++ libparsec/crates/types/src/pki/mod.rs
@@ -1,5 +1,7 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
+mod cert_ref;
+
 use std::{
     collections::HashMap,
     fmt::Display,
@@ -16,9 +18,12 @@ use libparsec_serialization_format::parsec_data;
 use crate::{
     self as libparsec_types, impl_transparent_data_format_conversion,
     serialization::{format_v0_dump, format_vx_load},
-    DataError, DataResult, DateTime, DeviceID, DeviceLabel, EnrollmentID, HumanHandle, ParsecAddr,
+    DataResult, DateTime, DeviceID, DeviceLabel, EnrollmentID, HumanHandle, ParsecAddr,
     ParsecPkiEnrollmentAddr, PkiEnrollmentLocalPendingError, PkiEnrollmentLocalPendingResult,
     UserID, UserProfile,
+};
+pub use cert_ref::{
+    X509CertificateHash, X509CertificateReference, X509CertificateReferenceIdOrHash,
 };
 
 /*
@@ -275,81 +280,6 @@ impl From<LocalPendingEnrollment> for LocalPendingEnrollmentData {
     }
 }
 
-#[derive(
-    Clone, Eq, PartialEq, serde_with::SerializeDisplay, serde_with::DeserializeFromStr, Debug,
-)]
-pub enum X509CertificateHash {
-    SHA256(Box<[u8; 32]>),
-}
-
-impl X509CertificateHash {
-    pub fn fake_sha256() -> Self {
-        Self::SHA256(Default::default())
-    }
-}
-
-impl std::fmt::Display for X509CertificateHash {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let (hash_str, data) = match self {
-            X509CertificateHash::SHA256(data) => ("sha256", data.as_ref()),
-        };
-        write!(
-            f,
-            "{hash_str}-{}",
-            ::data_encoding::BASE64.encode_display(data)
-        )
-    }
-}
-
-impl FromStr for X509CertificateHash {
-    type Err = DataError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let (hash_ty, b64_hash) = s.split_once('-').ok_or(DataError::BadSerialization {
-            format: None,
-            step: "Missing `-` delimiter",
-        })?;
-        let raw_data = data_encoding::BASE64
-            .decode(b64_hash.as_bytes())
-            .map_err(|_| DataError::BadSerialization {
-                format: None,
-                step: "error decoding hash",
-            })?;
-        if hash_ty.eq_ignore_ascii_case("sha256") {
-            Ok(X509CertificateHash::SHA256(raw_data.try_into().map_err(
-                |_| DataError::BadSerialization {
-                    format: None,
-                    step: "Invalid data size",
-                },
-            )?))
-        } else {
-            Err(DataError::BadSerialization {
-                format: None,
-                step: "Unsupported hash type ",
-            })
-        }
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub enum X509CertificateReference {
-    Id(Bytes),
-    Hash(X509CertificateHash),
-    IdOrHash(X509CertificateReferenceIdOrHash),
-}
-
-impl From<X509CertificateReferenceIdOrHash> for X509CertificateReference {
-    fn from(value: X509CertificateReferenceIdOrHash) -> Self {
-        Self::IdOrHash(value)
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct X509CertificateReferenceIdOrHash {
-    pub id: Bytes,
-    pub hash: X509CertificateHash,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum EncryptionAlgorithm {
     RsaesOaepSha256,
@@ -381,6 +311,6 @@ impl Display for EncryptionAlgorithm {
 }
 
 #[cfg(test)]
-#[path = "../tests/unit/pki.rs"]
+#[path = "../../tests/unit/pki.rs"]
 #[allow(clippy::unwrap_used)]
 mod tests;
